refactor(games): derive game availability from server-side feature flags
Split games page into server component (page.tsx) + client component (GamesClient.tsx).
Server reads feature flags at request time and passes enabledExperimental[] as a prop,
so experimental games (alias, liars_party, etc.) are only shown as available when
the corresponding ENABLE_* env var is set — not hardcoded in the UI.
